[TASK] Add new component for link list with tag filter
