+++ resources/js/Components/LinkList/LinkList.svelte
@@ -6,17 +6,16 @@
     import DropdownItem from "@/Components/Dropdowns/DropdownItem.svelte";
     import Dropdown from "@/Components/Dropdowns/Dropdown.svelte";
     import InnerDropdownSection from "@/Components/Dropdowns/InnerDropdownSection.svelte";
-    import {linkFilter, selectedTags} from "@/stores.js";
-    import {beforeUpdate, onDestroy} from "svelte";
+    import {selectedTags} from "@/stores.js";
     import {debounce} from "lodash";
     import Button from "@/Components/Buttons/Button.svelte";
     import Modal from "@/Components/Modals/Modal.svelte";
+    import {createEventDispatcher, onDestroy} from "svelte";
 
     export let links = [];
     export let searchString = '';
-    export let filteredTags = [];
-    export let showUntaggedOnly = false;
-    export let showTagFilter = true;
+
+    const dispatch = createEventDispatcher();
 
     let groupSelectMenu;
 
@@ -79,14 +78,8 @@
         }
     }
 
-    onDestroy(() => {
-        selectedTags.reset();
-    });
-
-    $: $linkFilter.isActive && filterLinks();
-
     const search = debounce(() => {
-        filterLinks();
+        dispatch('searched', searchString);
     }, 400);
 
     const clearSearchInput = () => {
@@ -94,37 +87,13 @@
         search();
     }
 
-    function filterLinks() {
-        if (searchString === '') {
-            searchString = null;
-        }
-
-        router.get(route(route().current(), {
-            group: route().params.group,
-            tags: $linkFilter.tags.length ? $linkFilter.tags.map(item => item.name).join(',') : null,
-            search: searchString,
-            untaggedOnly: $linkFilter.showUntaggedOnly ? true : null,
-        }), {}, {
-            only: ['links', 'searchString', 'filteredTags', 'showUntaggedOnly'],
-            preserveState: true,
-        });
-    }
-
-    function removeFilteredTag(tag) {
-        $linkFilter.tags = $linkFilter.tags.filter(item => item !== tag);
-    }
-
-    beforeUpdate(() => {
-        $linkFilter.tags = filteredTags;
-        $linkFilter.showUntaggedOnly = showUntaggedOnly;
-    });
-
     onDestroy(() => {
-        linkFilter.reset();
+        selectedTags.reset();
     });
 </script>
 
 <div class="px-4 sm:px-0">
+    <!-- Search input -->
     <div class="flex flex-col sm:flex-row">
         <div class="w-full sm:flex sm:items-center">
             <div class="relative rounded-md shadow-sm">
@@ -147,42 +116,11 @@
                 </button>
             {/if}
 
-            {#if showTagFilter}
-                <Button on:clicked={() => dispatchCustomEvent('filterTags')}
-                        title="Filter by tags" color="white" class="mt-4 sm:mt-0 sm:ml-6 sm:w-auto">
-                    <svg slot="icon" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" class="fill-gray-600">
-                        <path fill-rule="evenodd"
-                              d="M5.5 3A2.5 2.5 0 003 5.5v2.879a2.5 2.5 0 00.732 1.767l6.5 6.5a2.5 2.5 0 003.536 0l2.878-2.878a2.5 2.5 0 000-3.536l-6.5-6.5A2.5 2.5 0 008.38 3H5.5zM6 7a1 1 0 100-2 1 1 0 000 2z"
-                              clip-rule="evenodd"/>
-                    </svg>
-                </Button>
-                {#if $linkFilter.showUntaggedOnly}
-                    <button on:click={() => {$linkFilter.showUntaggedOnly = false; $linkFilter.isActive = true}}
-                            type="button"
-                            class="block mt-2 mx-auto text-sm text-gray-700 sm:mt-0 sm:mr-0 sm:ml-2">
-                        Show all
-                    </button>
-                {/if}
-            {/if}
+            <slot name="toolbar"/>
         </div>
     </div>
 
-    <!-- Tag filter list -->
-    {#if showTagFilter && $linkFilter.tags.length}
-        <div class="mt-4">
-            {#each $linkFilter.tags as tag (tag.id)}
-                <button on:click={() => {removeFilteredTag(tag); $linkFilter.isActive = true}} type="button"
-                        class="inline-flex items-center mr-2 mt-2 py-0.5 px-2.5 bg-primary-100 text-sm text-primary-800 font-medium rounded-full group">
-                    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor"
-                         class="mr-1 -ml-1 size-4 fill-primary-500 rounded-full group-hover:fill-primary-700 group-hover:bg-primary-200/50">
-                        <path
-                            d="M6.28 5.22a.75.75 0 00-1.06 1.06L8.94 10l-3.72 3.72a.75.75 0 101.06 1.06L10 11.06l3.72 3.72a.75.75 0 101.06-1.06L11.06 10l3.72-3.72a.75.75 0 00-1.06-1.06L10 8.94 6.28 5.22z"/>
-                    </svg>
-                    {tag.name}
-                </button>
-            {/each}
-        </div>
-    {/if}
+    <slot/>
 
     <!-- Edit links -->
     {#if links.data.length}
@@ -276,6 +214,7 @@
     {/if}
 </div>
 
+<!-- Link list -->
 <ul class="grid grid-cols-1 mt-2 divide-y divide-gray-200 sm:grid-cols-2 sm:gap-3 sm:mt-4 sm:divide-none">
     {#each links.data as link (link.id)}
         <li class="flex bg-white shadow sm:overflow-hidden sm:rounded-lg">
