+++ resources/js/Pages/Inbox/Index.svelte
@@ -9,18 +9,17 @@
     import LinkList from "@/Components/LinkList/LinkList.svelte";
     import {router} from "@inertiajs/svelte";
     import {route} from "@/utils/index.js";
-    import {linkFilter} from "@/stores.js";
 
     export let links = [];
-    export let searchString = '';
 
     let untagged = false;
     let ungrouped = false;
 
     $title = 'Inbox';
 
-    function filterLinks() {
-        router.post(route(route().current(), {}), {
+    function filterLinks(searchString = '') {
+        router.post(route(route().current()), {
+            search: searchString,
             untagged: untagged,
             ungrouped: ungrouped,
         }, {
@@ -34,5 +33,5 @@
     <button type="button" on:click={() => {untagged = !untagged; filterLinks()}}>Show untagged</button>
     <button type="button" on:click={() => {untagged = !untagged; filterLinks()}}>Show ungrouped</button>
 
-    <LinkList {links} {searchString} showTagFilter={false}/>
+    <LinkList on:searched={(e) => filterLinks(e.detail)} {links}/>
 </Main>
